+++ src/index.custom.js
@@ -1,7 +1,7 @@
 
 'use strict';
 
-var $ = require('jquery');
+var extend = require('extend');
 var config = require('./config');
 var setEnv = require('./set_env');
 var getEnv = require('./get_env');
@@ -9,8 +9,7 @@ var send = require('./send');
 var set = require('./set');
 
 set({
-    request: $.ajax,
-    extend: $.extend
+    extend: extend
 });
 
 var seeAjax = send;
@@ -18,5 +17,6 @@ var seeAjax = send;
 seeAjax.config = config;
 seeAjax.setEnv = setEnv;
 seeAjax.getEnv = getEnv;
+seeAjax.set = set;
 
 module.exports = seeAjax;