+++ src/setting.js
@@ -1,0 +1,7 @@
+
+module.exports = {
+    // custom request function
+    request: void 0,
+    // custom extend function
+    extend: void 0
+};